Move XML/JSON transformations into cyclestreets-core library
