+++ libraries/cyclestreets-core/build.gradle
@@ -23,13 +23,19 @@ dependencies {
 
   // Jackson is already included transitively from multiple sources; we specify explicit versions
   // to ensure consistency across these artifacts.
-  api 'com.fasterxml.jackson.core:jackson-annotations:2.9.5'
-  api 'com.fasterxml.jackson.core:jackson-core:2.9.5'
-  api 'com.fasterxml.jackson.core:jackson-databind:2.9.5'
+  implementation 'com.fasterxml.jackson.core:jackson-annotations:2.9.5'
+  implementation 'com.fasterxml.jackson.core:jackson-core:2.9.5'
+  implementation 'com.fasterxml.jackson.core:jackson-databind:2.9.5'
+
+  // XML / JSON conversion utilities
+  implementation 'com.github.smart-fun:XmlToJson:1.4.4'
+  implementation 'com.bazaarvoice.jolt:jolt-core:0.1.1'
+  implementation 'com.bazaarvoice.jolt:json-utils:0.1.1'
 
   testImplementation 'com.github.tomakehurst:wiremock-standalone:2.18.0'
   testImplementation 'junit:junit:4.12'
   testImplementation 'org.hamcrest:hamcrest-all:1.3'
+  testImplementation 'org.skyscreamer:jsonassert:1.5.0'
   testImplementation 'org.mockito:mockito-core:2.18.3'
   testImplementation 'org.robolectric:robolectric:3.3.2'
   testImplementation 'commons-io:commons-io:2.6'
