+++ libraries/cyclestreets-core/src/main/java/net/cyclestreets/api/client/JourneyStringTransformer.java
@@ -1,4 +1,4 @@
-package net.cyclestreets.content;
+package net.cyclestreets.api.client;
 
 import com.bazaarvoice.jolt.Chainr;
 import com.bazaarvoice.jolt.JsonUtils;
