+++ libraries/cyclestreets-core/src/test/resources/journey-domain.json
